Redo units.py and add support for relative units
Instead of data being handled directly by the string form of it, it is
split up into a unit (str) and data (float), reducing the need to
extract data and it's unit from a string.

A convert() method has been added to the Value class that converts it to
any other unit.

Support for relative units has been added, however relative units
require context, so a **context parameter has been added to
Value.convert(), get_px(), from_px(), convert(), and convert_str(). As
well as a Value.context attribute that is used in convert() if there is
not any context passed into it.

Conversion constants have been scrapped, in favour of a single function
(convert()) that can convert between any unit with given context.
from_px() and to_px() functions have been created which convert any unit
to and from pixels and are used in the convert() function.

Value.__init__ and convert() now only accept data in the form of a float
and the unit (str) in seperate parameters. There is, however a new
convert_str() function which takes data in the form of '<data><unit>'
and converts it to a specified unit with necessary context.